2.1.StreamLit/requirements.txt
streamlit
pillow
numpy
matplotlib
groq
